rename config to peekconfig
